[COMMUNICATION] Fix the timeout of the communication

+++ Core/Src/utils.c
@@ -3,7 +3,7 @@
 
 uint8_t computeCheckSum(uint8_t* data, uint32_t size)
 {
-	uint8_t sum = 0;
+	uint32_t sum = 0;
 	for (int i = 0; i < size ; i ++)
 	{
 		sum += data[i];
